backend/requirements.txt
fastapi
pydantic
uvicorn
langchain
langchain_experimental
langchain-community
langchain-core
langchain-google-genai
python-dotenv
tiktoken
faiss-cpu
streamlit
protobuf
mysql-connector-python
sentence-transformers
chromadb